There was a confusion with one of the plugins, which has been moved to a different package.
