+++ bundles/oms-payone-error/src/FondOfKudu/Zed/OmsPayoneError/Communication/Plugin/Oms/Condition/ErrorOrderAddressConditionPlugin.php
@@ -7,9 +7,7 @@ use Spryker\Zed\Kernel\Communication\AbstractPlugin;
 use Spryker\Zed\Oms\Dependency\Plugin\Condition\ConditionInterface;
 
 /**
- * @method \FondOfKudu\Zed\OmsPayoneError\OmsPayoneErrorConfig getConfig()
  * @method \FondOfKudu\Zed\OmsPayoneError\Persistence\OmsPayoneErrorRepositoryInterface getRepository()
- * @method \FondOfKudu\Zed\OmsPayoneError\Communication\OmsPayoneErrorCommunicationFactory getFactory()
  */
 class ErrorOrderAddressConditionPlugin extends AbstractPlugin implements ConditionInterface
 {
