+++ bundles/oms-payone-error/src/FondOfKudu/Zed/OmsPayoneError/Communication/Plugin/Oms/Condition/ErrorPayoneSettingsIncorrectConditionPlugin.php
@@ -7,8 +7,6 @@ use Spryker\Zed\Kernel\Communication\AbstractPlugin;
 use Spryker\Zed\Oms\Dependency\Plugin\Condition\ConditionInterface;
 
 /**
- * @method \FondOfKudu\Zed\OmsPayoneError\OmsPayoneErrorConfig getConfig()
- * @method \FondOfKudu\Zed\OmsPayoneError\Communication\OmsPayoneErrorCommunicationFactory getFactory()
  * @method \FondOfKudu\Zed\OmsPayoneError\Persistence\OmsPayoneErrorRepositoryInterface getRepository()
  */
 class ErrorPayoneSettingsIncorrectConditionPlugin extends AbstractPlugin implements ConditionInterface
