+++ bundles/oms-payone-error/src/FondOfKudu/Zed/OmsPayoneError/OmsPayoneErrorDependencyProvider.php
@@ -2,8 +2,6 @@
 
 namespace FondOfKudu\Zed\OmsPayoneError;
 
-use FondOfKudu\Zed\OmsPayoneError\Dependency\Facade\OmsPayoneErrorToPayoneFacadeBridge;
-use FondOfKudu\Zed\OmsPayoneError\Dependency\Facade\OmsPayoneErrorToPayoneFacadeInterface;
 use Orm\Zed\Payone\Persistence\SpyPaymentPayoneApiLogQuery;
 use Spryker\Zed\Kernel\AbstractBundleDependencyProvider;
 use Spryker\Zed\Kernel\Container;
@@ -14,24 +12,6 @@ class OmsPayoneErrorDependencyProvider extends AbstractBundleDependencyProvider
      * @var string
      */
     public const PROPEL_QUERY_PAYMENT_PAYONE_API_LOG = 'PROPEL_QUERY_PAYMENT_PAYONE_API_LOG';
-
-    /**
-     * @var string
-     */
-    public const FACADE_PAYONE = 'FACADE_PAYONE';
-
-    /**
-     * @param \Spryker\Zed\Kernel\Container $container
-     *
-     * @return \Spryker\Zed\Kernel\Container
-     */
-    public function provideCommunicationLayerDependencies(Container $container): Container
-    {
-        $container = parent::provideCommunicationLayerDependencies($container);
-        $container = $this->addPayoneFacade($container);
-
-        return $container;
-    }
 
     /**
      * @param \Spryker\Zed\Kernel\Container $container
@@ -54,22 +34,6 @@ class OmsPayoneErrorDependencyProvider extends AbstractBundleDependencyProvider
     protected function addPaymentPayoneApiLogQuery(Container $container): Container
     {
         $container[static::PROPEL_QUERY_PAYMENT_PAYONE_API_LOG] = static fn (): SpyPaymentPayoneApiLogQuery => SpyPaymentPayoneApiLogQuery::create();
-
-        return $container;
-    }
-
-    /**
-     * @param \Spryker\Zed\Kernel\Container $container
-     *
-     * @return \Spryker\Zed\Kernel\Container
-     */
-    protected function addPayoneFacade(Container $container): Container
-    {
-        $container[static::FACADE_PAYONE] = static fn (
-            Container $container
-        ): OmsPayoneErrorToPayoneFacadeInterface => new OmsPayoneErrorToPayoneFacadeBridge(
-            $container->getLocator()->payone()->facade(),
-        );
 
         return $container;
     }
