+++ bundles/oms-payone-error/src/FondOfKudu/Zed/OmsPayoneError/Persistence/OmsPayoneErrorPersistenceFactory.php
@@ -7,8 +7,6 @@ use Orm\Zed\Payone\Persistence\SpyPaymentPayoneApiLogQuery;
 use Spryker\Zed\Kernel\Persistence\AbstractPersistenceFactory;
 
 /**
- * @method \FondOfKudu\Zed\OmsPayoneError\OmsPayoneErrorConfig getConfig()
- * @method \FondOfKudu\Zed\OmsPayoneError\Persistence\OmsPayoneErrorQueryContainerInterface getQueryContainer()
  * @method \FondOfKudu\Zed\OmsPayoneError\Persistence\OmsPayoneErrorRepositoryInterface getRepository()
  */
 class OmsPayoneErrorPersistenceFactory extends AbstractPersistenceFactory
