+++ bundles/oms-retry-capture-process/src/FondOfKudu/Shared/OmsRetryCaptureProcess/Transfer/oms_retry_capture_process.transfer.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<transfers xmlns="spryker:transfer-01"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="spryker:transfer-01 http://static.spryker.com/transfer-01.xsd">
+
+    <transfer name="Order">
+        <property name="email" type="string"/>
+        <property name="customer" type="Customer"/>
+    </transfer>
+
+    <transfer name="Customer">
+        <property name="email" type="string" />
+    </transfer>
+</transfers>
