+++ bundles/oms-retry-capture-process/src/FondOfKudu/Zed/OmsRetryCaptureProcess/Communication/Plugin/Condition/OrderCreatedInsideTimelimitConditionPlugin.php
@@ -10,6 +10,7 @@ use Spryker\Zed\Oms\Dependency\Plugin\Condition\ConditionInterface;
 
 /**
  * @method \FondOfKudu\Zed\OmsRetryCaptureProcess\OmsRetryCaptureProcessConfig getConfig()
+ * @method \FondOfKudu\Zed\OmsRetryCaptureProcess\Communication\OmsRetryCaptureProcessCommunicationFactory getFactory()
  */
 class OrderCreatedInsideTimelimitConditionPlugin extends AbstractPlugin implements ConditionInterface
 {
