+++ bundles/oms-retry-capture-process/src/FondOfKudu/Zed/OmsRetryCaptureProcess/Dependency/Facade/OmsRetryCatpureProcessToPayoneFacadeInterface.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace FondOfKudu\Zed\OmsPayoneError\Dependency\Facade;
+namespace FondOfKudu\Zed\OmsRetryCaptureProcess\Dependency\Facade;
 
 use Generated\Shared\Transfer\OrderTransfer;
 
-interface OmsPayoneErrorToPayoneFacadeInterface
+interface OmsRetryCatpureProcessToPayoneFacadeInterface
 {
     /**
      * @param \Generated\Shared\Transfer\OrderTransfer $orderTransfer
