+++ dandelion.json
@@ -44,11 +44,11 @@
     },
     "oms-payone-error": {
       "path": "bundles/oms-payone-error",
-      "version": "1.0.0"
+      "version": "1.0.1"
     },
     "oms-retry-capture-process": {
       "path": "bundles/oms-retry-capture-process",
-      "version": "1.0.1"
+      "version": "1.0.2"
     }
   },
   "vcs": {
